backend/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
sentence-transformers==2.7.0
google-generativeai==0.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
uvicorn==0.24.0